backend/requirements.txt
Flask==2.3.6
Flask-RESTful==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Cors==3.0.10
PyJWT==2.8.0
mysqlclient==2.2.0
scikit-learn==1.3.3
pandas==2.2.3
numpy==1.26.0
